feat(client): clean up
