+++ packages/amplication-client/src/Workspaces/queries/projectQueries.ts
@@ -1,18 +1,5 @@
 import { gql } from "@apollo/client";
 
-export const GET_CURRENT_PROJECT = gql`
-  query findProject($projectId: String!) {
-    project(where: { id: $projectId }) {
-      id
-      name
-      resources {
-        id
-      }
-      createdAt
-    }
-  }
-`;
-
 export const GET_PROJECTS = gql`
   query findProjects {
     projects {
