+++ packages/amplication-client/src/Workspaces/queries/resourcesQueries.ts
@@ -1,5 +1,21 @@
 import { gql } from "@apollo/client";
 
+export const GET_RESOURCE = gql`
+  query getResource($id: String!) {
+    resource(where: { id: $id }) {
+      id
+      createdAt
+      updatedAt
+      name
+      description
+      githubLastSync
+      githubLastMessage
+      resourceType
+      isUnderLimitation
+    }
+  }
+`;
+
 export const GET_RESOURCES = gql`
   query getResources($projectId: String!, $whereName: StringFilter) {
     resources(
